# Speedup bincount and histc on CUDA